fix: greater index in broadcast

+++ sources/broadcast.c
@@ -6,7 +6,7 @@ void	broadcast(t_server *server, const int16_t except)
 {
 	const uint64_t	message_length = strlen(server->buffer);
 
-	for (uint8_t fd = 0; fd <= server->max_fd; ++fd)
+	for (uint64_t fd = 0; fd <= server->max_fd; ++fd)
 	{
 		if (fd ^ server->socket && fd ^ except && FD_ISSET(fd, &server->active))
 			send(fd, server->buffer, message_length, MSG_NOSIGNAL);
